backend/requirements.txt
pandas
confluent-kafka
python-dotenv
psycopg2-binary
fastapi
uvicorn[standard]
python-multipart